Open Source LLMs Path/FastMCP Foundations/requirements.txt
fastmcp>=0.2.1
requests>=2.31.0
python-dotenv>=1.0.1
